separate client dataclasses
