+++ src/sentry/taskworker/client/processing_result.py
@@ -1,0 +1,13 @@
+import dataclasses
+
+from sentry_protos.taskbroker.v1.taskbroker_pb2 import TaskActivationStatus
+
+
+@dataclasses.dataclass
+class ProcessingResult:
+    """Result structure from child processess to parent"""
+
+    task_id: str
+    status: TaskActivationStatus.ValueType
+    host: str
+    receive_timestamp: float
